+++ src/env.d.ts
@@ -5,3 +5,11 @@ declare namespace App {
     translate: (key: string, param?: string | number) => string
   }
 }
+
+interface ImportMetaEnv {
+  readonly PUBLIC_GOOGLE_ANALYTICS_ID: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}